requirements.txt
langchain
openai
tabulate
python-environ
notebook
kaggle
yt_dlp
ffmpeg
tiktoken
lark
pydub  # Load text with Whisper from videos
pytube  # Get videos from playlist video
streamlit
pydantic==1.10.8 # https://github.com/hwchase17/langchain/issues/7548
python-dotenv

# Vector DB
pinecone-client
chromadb==0.3.26
clickhouse-connect  # MyScale

# Confluence
atlassian-python-api
lxml
markdownify


gpt4all
